feat: oa sync

+++ api/src/main/java/com/hbcy/authcenter/api/modules/intgr/oa/vo/OaPageSizeModifyVO.java
@@ -12,5 +12,5 @@ import lombok.experimental.Accessors;
 @Accessors(chain = true)
 public class OaPageSizeModifyVO {
     private String dataKey;
-    private int pageSize;
+    private long pageSize;
 }
